requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==41.0.4
Deprecated==1.2.13
idna==3.4
pycparser==2.21
PyGithub==2.2.0
PyJWT==2.6.0
PyNaCl==1.5.0
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
typing_extensions==4.8.0
tzdata==2022.6
urllib3==1.26.12
wrapt==1.14.1
